Update: updated how app is generated based on comments
