+++ canvas_modules/create-elyra-canvas-app/template/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "{{APP_NAME}}",
+  "version": "0.1.0",
+  "type": "module",
+  "description": "{{APP_NAME}} - an Elyra Canvas application",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@carbon/icons-react": "^11.78.0",
+    "@carbon/react": "^1.105.0",
+    "@elyra/canvas": "^13.46.1",
+    "@ibm/plex-mono": "^1.1.0",
+    "@ibm/plex-sans": "^1.1.0",
+    "@ibm/plex-sans-condensed": "^2.0.0",
+    "@ibm/plex-serif": "^2.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-intl": "^5.25.1"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.3.0",
+    "sass": "^1.99.0",
+    "vite": "^5.0.0"
+  }
+}
