+++ canvas_modules/create-elyra-canvas-app/template/vite.config.js
@@ -1,0 +1,14 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+export default defineConfig({
+  plugins: [react()],
+  css: {
+    preprocessorOptions: {
+      scss: {
+        api: "modern-compiler",
+        loadPaths: ["node_modules"],
+      },
+    },
+  },
+});
